requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
drf-yasg
dj-config-url
psycopg2-binary
pillow
django-jazzmin
celery[redis]
django-filter
django-cors-headers
pyngrok==5.0.5
bs4
requests
lxml
selenium
webdriver_manager
fake-useragent
python-json-logger
twilio
django-credit-cards
python-dateutil
